requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
Flask==1.1.1
fsspec==0.6.2
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
numpydoc==0.9.2
pexpect==4.8.0
Pillow==7.0.0
requests==2.22.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.28.0
zict==1.0.0
zipp==2.2.0
